+++ lib/dry/validation.rb
@@ -1,32 +1,16 @@
 require 'dry-equalizer'
 require 'dry-configurable'
 require 'dry-container'
+require 'dry/core/extensions'
 
 require 'dry/validation/schema'
 require 'dry/validation/schema/form'
 require 'dry/validation/schema/json'
 
 module Dry
-  # FIXME: move this to dry-logic if it works lol
-  require 'dry/logic/predicate'
-  module Logic
-    class Predicate
-      class Curried < Predicate
-        def evaluate_args!(schema)
-          @args = args.map { |arg|
-            arg.is_a?(UnboundMethod) ? arg.bind(schema).() : arg
-          }
-          self
-        end
-      end
-
-      def evaluate_args!(*)
-        self
-      end
-    end
-  end
-
   module Validation
+    extend Dry::Core::Extensions
+
     MissingMessageError = Class.new(StandardError)
     InvalidSchemaError = Class.new(StandardError)
 
@@ -45,12 +29,16 @@ module Dry
       end
     end
 
-    def self.Form(options = {}, &block)
-      Validation.Schema(Schema::Form, options, &block)
+    def self.Form(base = nil, **options, &block)
+      klass = base ? Schema::Form.configure(Class.new(base)) : Schema::Form
+      Validation.Schema(klass, options, &block)
     end
 
-    def self.JSON(options = {}, &block)
-      Validation.Schema(Schema::JSON, options, &block)
+    def self.JSON(base = Schema::JSON, **options, &block)
+      klass = base ? Schema::JSON.configure(Class.new(base)) : Schema::JSON
+      Validation.Schema(klass, options, &block)
     end
   end
 end
+
+require 'dry/validation/extensions'