+++ lib/dry/validation/constants.rb
@@ -1,6 +1,0 @@
-module Dry
-  module Validation
-    EMPTY_ARRAY = [].freeze
-    EMPTY_HASH = {}.freeze
-  end
-end