+++ lib/dry/validation/message.rb
@@ -1,32 +1,55 @@
-require 'dry/validation/constants'
+require 'dry/core/constants'
 
 module Dry
   module Validation
     class Message
+      include Core::Constants
       include Dry::Equalizer(:predicate, :path, :text, :options)
-
-      Index = Class.new {
-        def inspect
-          "index"
-        end
-        alias_method :to_s, :inspect
-      }.new
 
       attr_reader :predicate, :path, :text, :rule, :args, :options
 
-      class Each < Message
-        def index_path
-          @index_path ||= [*path[0..path.size-2], Index]
+      class Or
+        attr_reader :left
+
+        attr_reader :right
+
+        attr_reader :path
+
+        attr_reader :messages
+
+        def initialize(left, right, messages)
+          @left = left
+          @right = right
+          @messages = messages
+          @path = left.path
         end
 
-        def each?
-          true
+        def hint?
+          false
+        end
+
+        def root?
+          path.empty?
+        end
+
+        def to_s
+          [left, right].uniq.join(" #{messages[:or]} ")
+        end
+      end
+
+      class Check < Message
+        def initialize(*args)
+          super
+          @path = [rule] unless rule.to_s.end_with?('?') || path.include?(rule)
         end
       end
 
       def self.[](predicate, path, text, options)
-        klass = options[:each] ? Message::Each : Message
-        klass.new(predicate, path, text, options)
+        if options[:check]
+          Message::Check.new(predicate, path, text, options)
+        else
+          Message.new(predicate, path, text, options)
+        end
       end
 
       def initialize(predicate, path, text, options)
@@ -35,22 +58,19 @@ module Dry
         @text = text
         @options = options
         @rule = options[:rule]
-        @each = options[:each] || false
         @args = options[:args] || EMPTY_ARRAY
-      end
 
-      alias_method :index_path, :path
+        if predicate == :key?
+          @path << rule
+        end
+      end
 
       def to_s
         text
       end
 
       def signature
-        @signature ||= [predicate, args, index_path].hash
-      end
-
-      def each?
-        @each
+        @signature ||= [predicate, args, path].hash
       end
 
       def hint?
@@ -68,14 +88,7 @@ module Dry
 
     class Hint < Message
       def self.[](predicate, path, text, options)
-        klass = options[:each] ? Hint::Each : Hint
-        klass.new(predicate, path, text, options)
-      end
-
-      class Each < Hint
-        def index_path
-          @index_path ||= [*path, Index]
-        end
+        Hint.new(predicate, path, text, options)
       end
 
       def hint?