+++ lib/dry/validation/predicate_registry.rb
@@ -1,6 +1,11 @@
+require 'dry/core/constants'
+require 'dry/logic/rule/predicate'
+
 module Dry
   module Validation
     class PredicateRegistry
+      include Core::Constants
+
       attr_reader :predicates
       attr_reader :external
 
@@ -23,10 +28,7 @@ module Dry
         end
 
         def update(other)
-          unbound_predicates = other.each_with_object({}) { |(n, p), res|
-            res[n] = Logic::Predicate.new(n, fn: p)
-          }
-          predicates.update(unbound_predicates)
+          predicates.update(other)
           self
         end
       end
@@ -59,7 +61,7 @@ module Dry
 
       def [](name)
         predicates.fetch(name) do
-          if external.key?(name)
+          if external.public_methods.include?(name)
             external[name]
           else
             raise_unknown_predicate_error(name)
@@ -68,7 +70,16 @@ module Dry
       end
 
       def key?(name)
-        predicates.key?(name) || external.key?(name)
+        predicates.key?(name) || external.public_methods.include?(name)
+      end
+
+      def arg_list(name, *values)
+        predicate = self[name]
+
+        predicate
+          .parameters
+          .map(&:last)
+          .zip(values + Array.new(predicate.arity - values.size, Undefined))
       end
 
       def ensure_valid_predicate(name, args_or_arity, schema = nil)