+++ lib/dry/validation/schema/dsl.rb
@@ -6,6 +6,7 @@ module Dry
     class Schema
       class DSL < BasicObject
         include ::Dry::Validation::Deprecations
+        include ::Dry::Core::Constants
 
         attr_reader :name, :registry, :rules, :checks, :parent, :options
 
@@ -15,6 +16,7 @@ module Dry
 
         def initialize(options = {})
           @name = options[:name]
+          @path = options.fetch(:path, name)
           @parent = options[:parent]
           @registry = options.fetch(:registry)
           @rules = options.fetch(:rules, [])
@@ -49,7 +51,7 @@ module Dry
         end
 
         def add_check(check)
-          checks << check
+          checks << check.to_rule
           self
         end
 
@@ -67,12 +69,16 @@ module Dry
         end
 
         def path
-          items = [parent && parent.path, name].flatten.compact.uniq
+          items = [parent && parent.path, *@path].flatten.compact.uniq
           items.size == 1 ? items[0] : items
         end
 
         def with(new_options)
           self.class.new(options.merge(new_options))
+        end
+
+        def predicate(name, args = EMPTY_ARRAY)
+          [:predicate, [name, registry.arg_list(name, *args)]]
         end
 
         def predicate?(meth)
@@ -104,7 +110,7 @@ module Dry
           end
         end
 
-        def create_rule(node)
+        def create_rule(node, name = self.name)
           Schema::Rule.new(node, name: name, target: self)
         end
       end