+++ spec/integration/schema/using_types_spec.rb
@@ -23,56 +23,29 @@ RSpec.describe Dry::Validation::Schema, 'defining schema using dry types' do
   end
 
   it 'passes when input is valid' do
-    expect(schema.(email: 'jane@doe', age: 19, country: 'Australia')).to be_success
-    expect(schema.(email: 'jane@doe', age: nil, country: 'Poland')).to be_success
+    expect(schema.call(email: 'jane@doe', age: 19, country: 'Australia')).to be_success
+    expect(schema.call(email: 'jane@doe', age: nil, country: 'Poland')).to be_success
   end
 
   it 'fails when input is not valid' do
-    expect(schema.(email: '', age: 19, country: 'New Zealand')).to_not be_success
-    expect(schema.(email: 'jane@doe', age: 17)).to_not be_success
-    expect(schema.(email: 'jane@doe', age: '19')).to_not be_success
+    expect(schema.call(email: '', age: 19, country: 'New Zealand')).to_not be_success
+    expect(schema.call(email: 'jane@doe', age: 17)).to_not be_success
+    expect(schema.call(email: 'jane@doe', age: '19')).to_not be_success
   end
 
   it 'correctly responds to messages' do
-    expect(schema.({}).messages).to eq(
-      age: ["is missing", "must be greater than 18"],
-      country: ["is missing", "must be one of: Australia, Poland"],
-      email: ["is missing", "must be String"],
+    expect(schema.call({}).messages).to eq(
+      age: ['is missing', 'must be Integer', 'must be greater than 18'],
+      country: ['is missing', 'must be String', 'must be one of: Australia, Poland'],
+      email: ['is missing', 'must be String']
     )
   end
 
   it 'fails when sum-type rule did not pass' do
-    result = schema.(email: 'jane@doe', age: 19, country: 'Australia', admin: 'foo')
+    result = schema.call(email: 'jane@doe', age: 19, country: 'Australia', admin: 'foo')
     expect(result.messages).to eql(
-      admin: ['must be FalseClass', 'must be TrueClass']
+      admin: ['must be TrueClass or must be FalseClass']
     )
-  end
-
-  context "structs" do
-    subject(:schema) do
-      Dry::Validation.Schema do
-        required(:person).filled(Person)
-      end
-    end
-
-    class Name < Dry::Types::Value
-      attribute :given_name, Dry::Types['strict.string']
-      attribute :family_name, Dry::Types['strict.string']
-    end
-
-    class Person < Dry::Types::Value
-      attribute :name, Name
-    end
-
-    it 'handles nested structs' do
-      expect(schema.(person: { name: { given_name: 'Tim', family_name: 'Cooper' } })).to be_success
-    end
-
-    it 'fails when input is not valid' do
-      expect(schema.(person: {name: {given_name: 'Tim'}}).messages).to eq(
-        person: { name: { family_name: ["is missing"] } }
-      )
-    end
   end
 
   context 'custom coercions' do
@@ -85,7 +58,7 @@ RSpec.describe Dry::Validation::Schema, 'defining schema using dry types' do
     end
 
     it 'applies custom types to input prior validation' do
-      result = schema.(email: ' jane@doe.org  ')
+      result = schema.call(email: ' jane@doe.org  ')
 
       expect(result).to be_success
       expect(result.to_h).to eql(email: 'jane@doe.org')
@@ -102,10 +75,46 @@ RSpec.describe Dry::Validation::Schema, 'defining schema using dry types' do
     end
 
     it 'applies custom types to input prior validation' do
-      result = schema.(quantity: '2', percentage: '0.5', switch: '0')
+      result = schema.call(quantity: '2', percentage: '0.5', switch: '0')
 
       expect(result).to be_success
       expect(result.to_h).to eql(quantity: 2, percentage: BigDecimal('0.5'), switch: false)
+    end
+  end
+
+  context 'with a nested schema' do
+    subject(:schema) do
+      Dry::Validation.Schema do
+        required(:address).schema do
+          zip = Dry::Types['strict.string'].constrained(format: /\A[0-9]{5}(-[0-9]{4})?\z/)
+
+          required(:zip).filled(zip)
+        end
+      end
+    end
+
+    it 'returns success for valid input' do
+      expect(schema.(address: { zip: '12321' })).to be_success
+    end
+
+    it 'returns failure for invalid input' do
+      expect(schema.(address: { zip: '12-321' })).to be_failure
+    end
+
+    it 'returns messages for invalid input' do
+      expect(schema.(address: nil).messages).to eql(
+        address: ['must be a hash']
+      )
+    end
+
+    it 'returns error messages for invalid input' do
+      expect(schema.(address: {}).errors).to eql(
+        address: { zip: ['is missing'] }
+      )
+
+      expect(schema.(address: { zip: '12-321' }).errors).to eql(
+        address: { zip: ['is in invalid format'] }
+      )
     end
   end
 
@@ -117,7 +126,7 @@ RSpec.describe Dry::Validation::Schema, 'defining schema using dry types' do
     end
 
     it 'applies type constraint checks to each element' do
-      result = schema.(countries: ['Poland', 'Australia'])
+      result = schema.call(countries: %w(Poland Australia))
 
       expect(result).to be_success
       expect(result.to_h).to eql(countries: %w(Poland Australia))