+++ spec/integration/schema/xor_spec.rb
@@ -13,8 +13,8 @@ RSpec.describe 'Schema with xor rules' do
 
       required(:have_cake).filled
 
-      rule(:be_reasonable) do
-        value(:eat_cake).eql?('yes!') ^ value(:have_cake).eql?('yes!')
+      rule(be_reasonable: %i[eat_cake have_cake]) do |ec, hc|
+        ec.eql?('yes!') ^ hc.eql?('yes!')
       end
     end
   end
@@ -27,9 +27,9 @@ RSpec.describe 'Schema with xor rules' do
     end
 
     it 'fails when both options are selected' do
-      messages = schema.(eat_cake: 'yes!', have_cake: 'yes!').messages[:be_reasonable]
-
-      expect(messages).to eql(['you cannot eat cake and have cake!'])
+      expect(schema.(eat_cake: 'yes!', have_cake: 'yes!').messages).to eql(
+        be_reasonable: ['you cannot eat cake and have cake!']
+      )
     end
   end
 end