+++ spec/spec_helper.rb
@@ -10,6 +10,7 @@ if RUBY_ENGINE == "rbx"
 end
 
 require 'dry-validation'
+require 'dry/core/constants'
 require 'ostruct'
 
 SPEC_ROOT = Pathname(__dir__)
@@ -18,6 +19,7 @@ Dir[SPEC_ROOT.join('shared/**/*.rb')].each(&method(:require))
 Dir[SPEC_ROOT.join('support/**/*.rb')].each(&method(:require))
 
 include Dry::Validation
+include Dry::Core::Constants
 
 module Types
   include Dry::Types.module
@@ -38,4 +40,17 @@ RSpec.configure do |config|
   end
 
   config.include PredicatesIntegration
+
+  config.before do
+    module Test
+      def self.remove_constants
+        constants.each { |const| remove_const(const)  }
+        self
+      end
+    end
+  end
+
+  config.after do
+    Object.send(:remove_const, Test.remove_constants.name)
+  end
 end